requirements.txt
flask==2.3.3
pandas==2.0.3
python-jobspy==1.1.55
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0